Following a Chrome/Firefox update, passive event listeners were added to addEventListener provoking many issues on event handling (i.e., e.preventDefalt() stopped to work).
In this library, it resulted on weird scroll behavior on mobile, offcanvas appearing / disappearing, console warnings and wroing handling of offcanvas status (open/close).

See this comment for a good overview of the problem: https://stackoverflow.com/a/39187679

For the time being, jQuery hasn't shipped a solution yet, so .on() is still unable to pass the parameters to the native addEventListener. See this issue to track the status of jQuery:  https://github.com/jquery/jquery/issues/2871

The only solution is to override touchmove event before loading jquery:
See https://stackoverflow.com/questions/39152877/consider-marking-event-handler-as-passive-to-make-the-page-more-responsive
and https://stackoverflow.com/questions/46094912/added-non-passive-event-listener-to-a-scroll-blocking-touchstart-event

The code has been inserted on bootstrap.offcanvas.coffee and needs to be loaded after jQuery and before the library.